Added support to deserialize CurrencyUnit (#118)
Currently when reading a property from config we have to use CurrencyCode enum because CurrencyUnit does not work. At the same time CurrencyUnit is the enum used to represent the currency of balances and amounts because it is embedded within Money.

By adding support to deserialize CurrencyUnit from Properties we are one step closer to be be able to use a single Enum in our code base.

Based on the fact that CurrencyUnit is embedded within Money and we won't be able to stop using Money; makes more sense to try to reduce the usage on CurrencyCode than CurrencyUnit.